requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
chromadb
pypdf
sentence-transformers
tavily-python
python-dotenv
tqdm
langgraph
openai
langchain-groq
langchain-chroma
langchain-core
